+++ src/lib/mongoBackend/safeBsonGet.h
@@ -50,7 +50,7 @@ extern int getIntField(const mongo::BSONObj& b, const std::string& field);
 *
 * getLongField -
 */
-extern long getLongField(const mongo::BSONObj& b, const std::string& field);
+extern long long getLongField(const mongo::BSONObj& b, const std::string& field);
 
 /* ****************************************************************************
 *
